Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.6
commonmark==0.9.1
cryptography==43.0.3
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
frozenlist==1.7.0
ggshield==1.43.0
git-filter-repo==2.47.0
idna==3.10
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
multidict==6.6.4
mypy_extensions==1.1.0
mysqlclient==2.2.7
numpy==2.3.3
oauthlib==3.2.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==3.0.0
propcache==0.3.2
pycparser==2.23
pygitguardian==1.25.0
Pygments==2.19.2
PyJWT==2.6.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.21.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
requests==2.32.5
rest-framework-simplejwt==0.0.2
rich==12.5.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
truststore==0.10.4
twilio==9.8.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
yarl==1.20.1
